formatted for better readability

+++ src/core/UniswapV2Factory.sol
@@ -7,8 +7,7 @@ import "./UniswapV2Pair.sol";
 contract UniswapV2Factory is IUniswapV2Factory {
     address public feeTo;
     address public feeToSetter;
-    bytes32 public constant INIT_CODE_PAIR_HASH =
-        keccak256(abi.encodePacked(type(UniswapV2Pair).creationCode));
+    bytes32 public constant INIT_CODE_PAIR_HASH = keccak256(abi.encodePacked(type(UniswapV2Pair).creationCode));
 
     mapping(address => mapping(address => address)) public getPair;
     address[] public allPairs;
@@ -17,18 +16,13 @@ contract UniswapV2Factory is IUniswapV2Factory {
         feeToSetter = _feeToSetter;
     }
 
-    function allPairsLength() external view returns (uint) {
+    function allPairsLength() external view returns (uint256) {
         return allPairs.length;
     }
 
-    function createPair(
-        address tokenA,
-        address tokenB
-    ) external returns (address pair) {
+    function createPair(address tokenA, address tokenB) external returns (address pair) {
         require(tokenA != tokenB, "UniswapV2:IDENTICAL_ADDRESS");
-        (address token0, address token1) = tokenA < tokenB
-            ? (tokenA, tokenB)
-            : (tokenB, tokenA); //to keep the order same ETH/USDC is same as USDC/ETH
+        (address token0, address token1) = tokenA < tokenB ? (tokenA, tokenB) : (tokenB, tokenA); //to keep the order same ETH/USDC is same as USDC/ETH
         require(token0 != address(0), "UniswapV2:ZERO_ADDRESS");
         require(getPair[token0][token1] == address(0), "UniswapV2:PAIR_EXISTS");
