+++ test/mocks/TestERC20Mock.sol
@@ -5,22 +5,18 @@ import {IERC20} from "../../src/core/interfaces/IERC20.sol";
 import {Context} from "../utils/Context.sol";
 import {IERC20Errors} from "../utils/IERC20Errors.sol";
 //Mock
+
 contract MockERC20 is IERC20, Context, IERC20Errors {
     mapping(address account => uint256) private _balances;
 
-    mapping(address account => mapping(address spender => uint256))
-        private _allowances;
+    mapping(address account => mapping(address spender => uint256)) private _allowances;
 
     uint256 private _totalSupply;
 
     string private _name;
     string private _symbol;
 
-    constructor(
-        string memory name_,
-        string memory symbol_,
-        uint _initialSupply
-    ) {
+    constructor(string memory name_, string memory symbol_, uint256 _initialSupply) {
         _name = name_;
         _symbol = symbol_;
         _mint(msg.sender, _initialSupply);
@@ -53,46 +49,28 @@ contract MockERC20 is IERC20, Context, IERC20Errors {
         return true;
     }
 
-    function allowance(
-        address owner,
-        address spender
-    ) public view virtual returns (uint256) {
+    function allowance(address owner, address spender) public view virtual returns (uint256) {
         return _allowances[owner][spender];
     }
 
-    function approve(
-        address spender,
-        uint256 value
-    ) public virtual returns (bool) {
+    function approve(address spender, uint256 value) public virtual returns (bool) {
         address owner = _msgSender();
         _approve(owner, spender, value);
         return true;
     }
 
-    function transferFrom(
-        address from,
-        address to,
-        uint256 value
-    ) public virtual returns (bool) {
+    function transferFrom(address from, address to, uint256 value) public virtual returns (bool) {
         address spender = _msgSender();
         _spendAllowance(from, spender, value);
         _transfer(from, to, value);
         return true;
     }
 
-    function _spendAllowance(
-        address owner,
-        address spender,
-        uint256 value
-    ) internal virtual {
+    function _spendAllowance(address owner, address spender, uint256 value) internal virtual {
         uint256 currentAllowance = allowance(owner, spender);
         if (currentAllowance < type(uint256).max) {
             if (currentAllowance < value) {
-                revert ERC20InsufficientAllowance(
-                    spender,
-                    currentAllowance,
-                    value
-                );
+                revert ERC20InsufficientAllowance(spender, currentAllowance, value);
             }
             unchecked {
                 _approve(owner, spender, currentAllowance - value, false);
@@ -144,12 +122,7 @@ contract MockERC20 is IERC20, Context, IERC20Errors {
         emit Transfer(from, to, value);
     }
 
-    function _approve(
-        address owner,
-        address spender,
-        uint256 value,
-        bool emitEvent
-    ) internal virtual {
+    function _approve(address owner, address spender, uint256 value, bool emitEvent) internal virtual {
         if (owner == address(0)) {
             revert ERC20InvalidApprover(address(0));
         }
